Added starring loops

+++ screens/editloop/index.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useCallback } from 'react';
-import { View, Text, Image, ActivityIndicator, Pressable, ScrollView, RefreshControl, TouchableOpacity, Touchable } from 'react-native';
+import { View, Text, Image, ActivityIndicator, Pressable, ScrollView, RefreshControl, TouchableOpacity, Switch, Alert } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 
 import { pickImage } from '../../components/imagepicker';
@@ -12,6 +12,19 @@ export default function EditLoop({ navigation, route }) {
     const [image_url, setImageURL] = useState(null)
     const [refreshing, setRefreshing] = useState(false)
 
+    const [membersAllowed, setMembersAllowed] = useState();
+    const [isPrivate, setIsPrivate] = useState()
+
+    async function toggleSwitch() {
+      setMembersAllowed(previousState => !previousState)
+      await updateLoop({ loop_id: loop.loop_id, endpoint: "allowMembersToCreateEvents", value: !loop.allowMembersToCreateEvents})
+    }
+
+    async function togglePrivate() {
+      setIsPrivate(previousState => !previousState)
+      await updateLoop({ loop_id: loop.loop_id, endpoint: "public", value: !loop.public})
+      await fetchLoop()
+    }
 
     async function handleImageSelect(image) {
         console.log("Selected Image in EditLoop:", image);
@@ -19,7 +32,7 @@ export default function EditLoop({ navigation, route }) {
         try {
             res = await uploadToCDN(image)
             console.log(res.url)
-            await updateLoop(loop_id, "pfp_url", res.url)
+            await updateLoop({loop_id: loop_id, endpoint: "pfp_url", value: res.url})
         } catch (error) {   
             console.error('Error uploading image:', error.message);
         }
@@ -35,19 +48,30 @@ export default function EditLoop({ navigation, route }) {
     
     
     async function handleDelete() {
-      await deleteLoop(loop_id)
-      navigation.navigate("Community")
+      Alert.alert("Are you sure you want to delete this Loop?", "This is a permanent action that cannot be undone.", [
+        {
+          text: 'Cancel',
+          onPress: () => console.log('Cancel Pressed'),
+          style: 'cancel',
+        },
+        {text: 'OK', onPress: async() => {
+            await deleteLoop(loop_id)
+            navigation.navigate("Community")}
+        },
+      ]);
     }
 
     async function fetchLoop() {
       const loop = await getLoop(loop_id)
       setLoop(loop)
       setImageURL(loop.pfp_url)
+      setMembersAllowed(loop.allowMembersToCreateEvents)
+      setIsPrivate(!loop.public)
     }
     
     useEffect(() => {
       fetchLoop()
-    }, []);
+    }, [route.params]);
 
     
     if (!loop) {
@@ -72,7 +96,7 @@ export default function EditLoop({ navigation, route }) {
                     </View>
                 )}
 
-                <Ionicons name="add-circle" size={32} color="blue" style={{position: "absolute", top: 0, right: 0}}/>
+                <Ionicons name="add-circle" size={32} color="rgb(47, 139, 128)" style={{position: "absolute", top: 0, right: 0}}/>
               </TouchableOpacity>
           </View>
             
@@ -90,7 +114,33 @@ export default function EditLoop({ navigation, route }) {
             <TouchableOpacity onPress={handleDelete} style={styles.updateField}>
               <Text style={{color: "red", flex: 0.5}}>Delete Loop</Text>
             </TouchableOpacity>
+          </View>
 
+          <View style={{paddingHorizontal: 15}}>
+              <View style={{ borderColor: "white", borderWidth: 1, padding: 15, flexDirection: "row", justifyContent: "space-between", borderRadius: 20, marginTop: 20}}>
+                <Text style={{color: "white", width: 150}}>Allow all members to create events</Text>
+
+                <Switch
+                    trackColor={{true: 'rgb(47, 139, 128)'}}
+                    onValueChange={toggleSwitch}
+                    value={membersAllowed}
+                    />
+              </View>
+          </View>
+
+          <View style={{paddingHorizontal: 15}}>
+              <View style={{ borderColor: "white", borderWidth: 1, padding: 15, flexDirection: "row", justifyContent: "space-between", borderRadius: 20, marginTop: 20}}>
+                <View>
+                  <Text style={{color: "white"}}>Private Loop</Text>
+                  <Text style={{color: "gray", fontSize: 10, width: 200, marginTop: 2}}>When your Loop is private, members have to request to join and only you can approve their request.</Text>
+                </View>
+                
+                <Switch
+                    trackColor={{true: 'rgb(47, 139, 128)'}}
+                    onValueChange={togglePrivate}
+                    value={isPrivate}
+                    />
+              </View>
           </View>
         </ScrollView>
     )
